pyproject.toml
[project]
name = "mako"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi[standard]>=0.135.3",
    "pwdlib[argon2]>=0.3.0",
    "pyjwt>=2.12.1",
]
